requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
httpx==0.27.0
pytest==8.2.0
